Rae language change: `Float` is now f32 (alias of Float32), not f64
Permanent language-design decision. Rae targets games, graphics, UI and
other performance-sensitive native software, where the common representation
for vectors, matrices, transforms, geometry, UI coordinates, colours,
vertices and particles is f32 — and where WGSL's ordinary float is f32.
Defaulting to f64 (the Python/JS convention) cost half the cache density and
bandwidth and imposed a routine f64->f32 narrowing on every GPU upload.
Rationale in full: docs/primitive-types.md.

    Float = Float32 = IEEE-754 binary32   (C float)
    Float64         = IEEE-754 binary64   (C double)

TYPE SYSTEM — Float32/Float64 are now REAL types, not backend-only strings.
Previously the front end had a single TYPE_FLOAT and the C backend decided
width by matching type-name strings; Float64/Float32 were never resolved in
sema at all. Now:
- TYPE_FLOAT is f32; new TYPE_FLOAT64 is a distinct kind (type.h/type.c).
- sema resolves `Float` and `Float32` to the SAME interned TypeInfo, so they
  share identity, equality, ABI, layout and monomorphization: List(Float)
  and List(Float32) are one instantiation by construction.
- `Float64` resolves separately, so List(Float64) is distinct.

MANGLING — fixes a latent collision: `Float` and `Float64` previously
mangled to the same name, which would have given List(Float) and
List(Float64) one monomorphization with DIFFERENT element sizes. Float/
Float32 now mangle as "Float", Float64 as "Float64". type.c had a second,
independent mangler that still emitted "double"; it is now aligned, which
was the cause of rae_List_float being defined but rae_List_double referenced.

C BACKEND — Float/Float32 -> `float`, Float64 -> `double`, emitted as plain
idiomatic C. No Rae-specific numeric typedefs were introduced.

FFI/ABI — the boundary is literal: Buffer(Float) is `float*`, Buffer(Float64)
is `double*`; there is no hidden widening. ~70 extern C signatures were
flipped in lockstep with their Rae declarations. The dangerous case (a Rae
float buffer read by C as double*) was verified to be caught by the C
compiler for header-declared externs; runtime/gpu externs not in the shared
header were audited by hand and by compiling generated C against the runtime.

GPU — removed the mesh-upload temp buffer in runtime_gpu3d.c that existed
solely to narrow f64->f32; vertex data now uploads directly. The i64->u32
index conversion is still semantically required and stays. The frame/draw
uniform repacking also stays: it reorders into padded std140 layout, which
is required independently of width.

INTENTIONAL Float64 — epoch timestamps. `gpu2d.nowSeconds()` and
`sys.fileMtime`/`fileModTime` return Float64 because a ~1.7e9 timestamp in
f32 has ~128s granularity: two calls one frame apart return the SAME value,
so every animation delta would be exactly 0.0 and file-watching would never
observe a change. hot_reload/file_watch `lastMtime` fields and locals were
widened to match; that chain was a real bug this change surfaced (415).

TESTS — new 550_float_is_f32 proves Float has f32 semantics, Float32 is the
same type, Float64 is distinct and retains f64 semantics, the alias is
interchangeable, and List(Float)/List(Float32) coexist. Uses 2^24 (last
integer f32 represents exactly) compared rather than printed, so it does not
depend on float formatting.

DOCS — new docs/primitive-types.md; corrected FFI.md, noise.md and
standard-library-ownership-model.md. noise.md now states that CPU and GPU
share a representation but does NOT claim bit-identical determinism:
operation order, FMA contraction and transcendental precision still differ.

+++ compiler/runtime/runtime_buffers_math.c
@@ -23,7 +23,7 @@ static uint32_t rae_next_u32(void) {
   return (uint32_t)(g_rae_random_state >> 32);
 }
 
-double rae_ext_rae_random(void) {
+float rae_ext_rae_random(void){
   return (double)rae_next_u32() / (double)4294967295.0;
 }
 
@@ -33,8 +33,8 @@ int64_t rae_ext_rae_random_int(int64_t min, int64_t max) {
   return min + (int64_t)(rae_next_u32() % range);
 }
 
-double rae_ext_rae_int_to_float(int64_t v) { return (double)v; }
-int64_t rae_ext_rae_float_to_int(double v) { return (int64_t)v; }
+float rae_ext_rae_int_to_float(int64_t v){ return (double)v; }
+int64_t rae_ext_rae_float_to_int(float v){ return (int64_t)v; }
 /* Debug-only bounds checking for rae_buf_get/set. Compiled in when the
  * binary is built with `-DRAE_DEBUG_BOUNDS`. Tracks (ptr -> count, elem_size)
  * in a small open-addressed hash; on every get/set the entry is looked up
@@ -214,20 +214,26 @@ RaeAny rae_ext_rae_buf_get_any(void* buf, int64_t index) {
   return ((RaeAny*)buf)[index];
 }
 
-double rae_ext_math_sin(double x) { return sin(x); }
-double rae_ext_math_cos(double x) { return cos(x); }
-double rae_ext_math_tan(double x) { return tan(x); }
-double rae_ext_math_asin(double x) { return asin(x); }
-double rae_ext_math_acos(double x) { return acos(x); }
-double rae_ext_math_atan(double x) { return atan(x); }
-double rae_ext_math_atan2(double y, double x) { return atan2(y, x); }
-double rae_ext_math_sqrt(double x) { return sqrt(x); }
-double rae_ext_math_pow(double base, double exp) { return pow(base, exp); }
-double rae_ext_math_exp(double x) { return exp(x); }
-double rae_ext_math_math_log(double x) { return log(x); }
-double rae_ext_math_floor(double x) { return floor(x); }
-double rae_ext_math_ceil(double x) { return ceil(x); }
-double rae_ext_math_round(double x) { return round(x); }
+/* lib/math.rae declares these over `Float`, which is f32 (see
+ * docs/primitive-types.md), so the C side takes and returns `float` and uses
+ * the f-suffixed libm entry points. Computing in float rather than
+ * round-tripping through double keeps the CPU result in the same
+ * representation the GPU uses, which is the point of an f32 default.
+ * Higher-precision variants belong on explicit Float64 APIs. */
+float rae_ext_math_sin(float x) { return sinf(x); }
+float rae_ext_math_cos(float x) { return cosf(x); }
+float rae_ext_math_tan(float x) { return tanf(x); }
+float rae_ext_math_asin(float x) { return asinf(x); }
+float rae_ext_math_acos(float x) { return acosf(x); }
+float rae_ext_math_atan(float x) { return atanf(x); }
+float rae_ext_math_atan2(float y, float x) { return atan2f(y, x); }
+float rae_ext_math_sqrt(float x) { return sqrtf(x); }
+float rae_ext_math_pow(float base, float exp) { return powf(base, exp); }
+float rae_ext_math_exp(float x) { return expf(x); }
+float rae_ext_math_math_log(float x) { return logf(x); }
+float rae_ext_math_floor(float x) { return floorf(x); }
+float rae_ext_math_ceil(float x) { return ceilf(x); }
+float rae_ext_math_round(float x) { return roundf(x); }
 
 /* JSON helpers for C backend */
 static const char* rae_json_find_key(const char* json, int64_t json_len, const char* key) {
