+++ compiler/runtime/runtime_webgpu.c
@@ -82,9 +82,9 @@ static int rae_wgpu_init(void) {
 
 /* scene: sceneLen f64 (camera 19 + spheres*10) -> narrowed to f32 for the GPU.
  * fb: width*height int64 written as packed 0xRRGGBB. wgsl: shader source. */
-void rae_ext_webgpu_raytrace(const double* scene, int64_t sceneLen, int64_t* fb,
+void rae_ext_webgpu_raytrace(const float* scene, int64_t sceneLen, int64_t* fb,
                             int64_t width, int64_t height, int64_t samples,
-                            int64_t maxDepth, rae_String wgsl) {
+                            int64_t maxDepth, rae_String wgsl){
     if (!fb || width <= 0 || height <= 0) return;
     if (!rae_wgpu_init()) return;
 
@@ -188,7 +188,7 @@ static int rae_gpu_add_buf(WGPUBuffer b, size_t size) {
     return ++g_gpu_buf_n;  /* 1-based handle */
 }
 
-int64_t rae_ext_gpu_storageF32(const double* data, int64_t count) {
+int64_t rae_ext_gpu_storageF32(const float* data, int64_t count){
     if (!rae_wgpu_init() || count <= 0) return 0;
     size_t bytes = (size_t)count * 4;
     float* tmp = (float*)malloc(bytes);
@@ -225,7 +225,7 @@ static int64_t rae_gpu_alloc(int64_t count) {
 int64_t rae_ext_gpu_allocF32(int64_t count) { return rae_gpu_alloc(count); }
 int64_t rae_ext_gpu_allocU32(int64_t count) { return rae_gpu_alloc(count); }
 
-void rae_ext_gpu_writeF32(int64_t buf, const double* data, int64_t count) {
+void rae_ext_gpu_writeF32(int64_t buf, const float* data, int64_t count){
     if (!g_wgpu_queue || buf < 1 || buf > g_gpu_buf_n || count <= 0) return;
     size_t bytes = (size_t)count * 4;
     float* tmp = (float*)malloc(bytes);
@@ -312,7 +312,7 @@ static const void* rae_gpu_readback(int64_t buf, size_t bytes, WGPUBuffer* stagi
     *staging_out = staging;
     return wgpuBufferGetConstMappedRange(staging, 0, bytes);
 }
-void rae_ext_gpu_downloadF32(int64_t buf, double* out, int64_t count) {
+void rae_ext_gpu_downloadF32(int64_t buf, float* out, int64_t count){
     if (!out || count <= 0) return;
     WGPUBuffer staging = NULL;
     const float* p = (const float*)rae_gpu_readback(buf, (size_t)count * 4, &staging);
